recipes-go/3mdeb-rtectrl/3mdeb-rtectrl/RteCtrl.cfg: Fix GPIO numbers
The GPIO numbering changed to dynamic in newer kernel. Now all GPIOs
from the expander start at offset 512 instead of 400. However, not
all GPIOs in the RteCtrl.cfg were updated correctly and some are even
duplicated. Update the GPIO numbers for duplicated entries so that
it matches the previous numbering based on offset 400.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>